requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
flask>=3.0.0
python-dateutil>=2.8.2
schedule>=1.2.1
markdown>=3.6
